+++ src/livelywpf/livelywpf/Core/Wallpapers/VideoPlayerVLCExt.cs
@@ -209,7 +209,7 @@ namespace livelywpf.Core
             SendMessage("lively:vid-volume " + volume);
         }
 
-        public void SetPlaybackPos(int pos)
+        public void SetPlaybackPos(float pos)
         {
             //todo
         }
